requirements.txt
streamlit
langchain_groq
langchain_core
langchain_community
requests
pandas
numpy
python-dotenv
google-generativeai==0.8.0
beautifulsoup4
